Wrong dtype

+++ DeepTreeAttention/generators/boxes.py
@@ -156,8 +156,8 @@ def generate_tfrecords(shapefile,
             chunk_labels = None
 
         #resize crops
-        resized_HSI_crops = [resize(x, HSI_size, HSI_size) for x in chunk_HSI_crops]
-        resized_RGB_crops = [resize(x, RGB_size, RGB_size) for x in chunk_RGB_crops]
+        resized_HSI_crops = [resize(x, HSI_size, HSI_size).astype(np.uint16) for x in chunk_HSI_crops]
+        resized_RGB_crops = [resize(x, RGB_size, RGB_size).astype(np.uint8) for x in chunk_RGB_crops]
 
         filename = "{}/{}_{}.tfrecord".format(savedir, basename, counter)
         
@@ -319,7 +319,7 @@ def _HSI_parse_(tfrecord):
     example = tf.io.parse_single_example(tfrecord, features)
 
     # Load HSI image from file
-    HSI_image = tf.io.decode_raw(example['HSI_image/data'], tf.uint8)
+    HSI_image = tf.io.decode_raw(example['HSI_image/data'], tf.uint16)
     HSI_image_shape = tf.stack([example['HSI_image/height'],example['HSI_image/width'], example['HSI_image/depth']])
     
     # Reshape to known shape
